backend/services/integrated_investment_service/server/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
yfinance>=0.2.36
pandas>=2.0.0
numpy>=1.24.0
google-genai>=1.0.0
openai>=1.0.0
anthropic>=0.39.0
python-dotenv>=1.0.0
mplfinance>=0.12.9b7
matplotlib>=3.8.0
Pillow>=10.0.0
google-cloud-storage>=2.0.0
